Combine betfair.meta.model.Model and betfair.model.BetfairModel
As only one type of model is used in this project (BetfairModel) it is
unnecessary for the Model base class to exist - no other classes need
to extend it. This subsumes the Model class' implementation into
BetfairModel.
