+++ betfair/meta/__init__.py
@@ -1,5 +1,4 @@
 from .datatype import DataType
 from .field import Field, ListField
-from .model import Model
 
-__all__ = ['DataType', 'Field', 'ListField', 'Model']
+__all__ = ['DataType', 'Field', 'ListField']
